requirements.txt
google-generativeai>=0.3.0
google-genai>=1.0.0
pillow>=10.0.0
